requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
filelock==3.15.4
Flask==3.0.3
fonttools==4.53.1
fsspec==2024.6.1
idna==3.8
imageio==2.35.1
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
lapx==0.5.10.post1
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
paho-mqtt==2.1.0
pandas==2.2.2
pillow==10.4.0
plotly==5.23.0
psutil==6.0.0
py-cpuinfo==9.0.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
retrying==1.3.4
scipy==1.14.1
seaborn==0.13.2
setuptools==73.0.1
shapely==2.0.6
six==1.16.0
sympy==1.13.2
tenacity==9.0.0
tifffile==2024.8.24
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.82
ultralytics-thop==2.0.5
urllib3==2.2.2
Werkzeug==3.0.4
zipp==3.20.1
